added the capability for each reader to be reused by the ReaderFactory.
This solution is a POC and very basic.

Every StructReader now has a reuse() method, which will put the Reader into a pool.
Whenever a new Reader is created by the factory it first checks if a Reader is available. if so, it removed this reader and inits it with the new values.

Attention: a Reader put into the pool is not usable until it is initialized again.